Fix Clang warning.

+++ src/baseq2/p_view.c
@@ -961,7 +961,7 @@ void ClientEndServerFrame(edict_t *ent)
         bobtime *= 4;
 
     bobcycle = (int)bobtime;
-    bobfracsin = fabsf(sin(bobtime * M_PI));
+    bobfracsin = fabs(sin(bobtime * M_PI));
 
     // detect hitting the floor
     P_FallingDamage(ent);
